feat: adding waffle + fixing up failing issues
