+++ constants.ts
@@ -5,5 +5,6 @@ export const MAX_SUPPLY = DECIMALS_MULTIPLIER.mul(500);
 export const NAME = "Nakamot-Os";
 export const SYMBOL = "BOX";
 export const NFT_URI = "https://arweave.net/w8Vs3I-TmRIcZVNaSkxjVhtWc9ZJleCJeSRlpXEVdPM";
-export const NFT_SUPPLY = 10;
+export const MAX_NFT_SUPPLY = 10;
 export const BLOCKS_TIL_LOTTO = 69420; // about 1.5 weeks
+export const NETWORK = "kovan";
