+++ hardhat.config.ts
@@ -10,7 +10,6 @@ import "hardhat-gas-reporter";
 import "hardhat-typechain";
 import "solidity-coverage";
 import "@nomiclabs/hardhat-ethers";
-import "hardhat-deploy-ethers";
 
 const accounts = {
     count: 10,
