+++ package.json
@@ -25,7 +25,7 @@
     "@types/mocha": "^7.0.2",
     "@types/node": "^14.11.8",
     "@typescript-eslint/eslint-plugin": "^4.14.0",
-    "chai": "^4.2.0",
+    "chai": "^4.3.4",
     "commitizen": "^4.2.1",
     "cz-conventional-changelog": "^3.3.0",
     "dotenv": "^8.2.0",
@@ -33,7 +33,7 @@
     "eslint-config-airbnb-base-typescript-prettier": "^4.1.0",
     "eslint-config-prettier": "^6.12.0",
     "eslint-plugin-import": "^2.22.1",
-    "ethereum-waffle": "^3.2.2",
+    "ethereum-waffle": "^3.3.0",
     "fs-extra": "^9.0.1",
     "hardhat": "^2.0.10",
     "hardhat-deploy": "^0.7.0-beta.45",
@@ -41,7 +41,7 @@
     "hardhat-gas-reporter": "^1.0.4",
     "hardhat-typechain": "^0.3.5",
     "husky": "^4.3.0",
-    "mocha": "^8.1.3",
+    "mocha": "^8.4.0",
     "prettier": "^2.1.2",
     "prettier-plugin-solidity": "^1.0.0-beta.1",
     "shelljs": "^0.8.4",
@@ -68,6 +68,7 @@
     "access": "public"
   },
   "scripts": {
+    "build": "waffle",
     "clean": "hardhat clean",
     "commit": "git-cz",
     "compile": "hardhat compile",
