+++ test/helpers/setup.ts
@@ -1,20 +1,37 @@
 import { deployments, ethers } from "hardhat";
 import { deploy } from ".";
-import { MAX_SUPPLY, NAME, SYMBOL, NFT_URI, MAX_NFT_SUPPLY } from "../../constants";
+import { MAX_SUPPLY, NAME, SYMBOL, NFT_URI, MAX_NFT_SUPPLY, NETWORK, BLOCKS_TIL_LOTTO } from "../../constants";
 import { NakamotOsERC20, NakamotOsERC721 } from "../../typechain";
+import { networkConfig, getNetworkIdFromName } from "../../config/networks";
+import { link } from "fs-extra";
 
 const setup = deployments.createFixture(async () => {
     const admin = await ethers.getNamedSigner("admin");
     const bagHolderAddress = await admin.getAddress();
-    const nft = (await deploy("NakamotOsERC721", {
+    const networkId = Number(getNetworkIdFromName(NETWORK));
+    const networkParams = networkConfig[networkId];
+    const { fee, keyHash, vrfCoordinator, linkToken } = networkParams;
+    const nft = ((await deploy("NakamotOsERC721", {
         args: [NAME, SYMBOL, NFT_URI],
         connect: admin,
         from: await admin.getAddress(),
-    })) as NakamotOsERC721;
-    const token = (await deploy("NakamotOsERC20", {
-        args: [NAME, SYMBOL, MAX_SUPPLY, await admin.getAddress(), nft.address, MAX_NFT_SUPPLY],
+    })) as unknown) as NakamotOsERC721;
+    const token = ((await deploy("NakamotOsERC20", {
+        args: [
+            NAME,
+            SYMBOL,
+            MAX_SUPPLY,
+            await admin.getAddress(),
+            nft.address,
+            MAX_NFT_SUPPLY,
+            BLOCKS_TIL_LOTTO,
+            keyHash,
+            vrfCoordinator,
+            linkToken,
+            fee,
+        ],
         connect: admin,
-    })) as NakamotOsERC20;
+    })) as unknown) as NakamotOsERC20;
 
     await nft.setERC20Address(token.address);
     await nft.renounceOwnership();
