+++ waffle.json
@@ -1,0 +1,6 @@
+{
+    "compilerType": "solcjs",
+    "compilerVersion": "0.6.9",
+    "sourceDirectory": "./contracts",
+    "outputDirectory": "./build"
+}
